Store the timestamp of the query, not just the order
This makes it easier to pick up when the session actually went
idle in some cases.

Intentionally store the text version of the timestamp, so we don't
call any more functions than necessary when dumping the query to
the log. This might add a couple of bytes of extra storage, but not
much.